+++ packages/yourloops/components/chat/chat-widget.tsx
@@ -45,7 +45,6 @@ import Box from '@mui/material/Box'
 import Select, { type SelectChangeEvent } from '@mui/material/Select'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
 import MenuItem from '@mui/material/MenuItem'
-import GenericDashboardCard from '../dashboard-widgets/generic-dashboard-card'
 import FormGroup from '@mui/material/FormGroup'
 import TeamUtils from '../../lib/team/team.util'
 import { getUserName } from '../../lib/auth/user.util'
@@ -77,7 +76,9 @@ const chatWidgetStyles = makeStyles({ name: 'ylp-chat-widget' })((theme: Theme) 
       }
     },
     chatWidgetContent: {
-      padding: theme.spacing(2),
+      paddingBottom: theme.spacing(2),
+      paddingLeft: theme.spacing(2),
+      paddingRight: theme.spacing(2),
       minHeight: CHAT_CONTENT_MIN_HEIGHT,
       maxHeight: CHAT_CONTENT_MAX_HEIGHT,
       overflowY: 'auto',
@@ -85,7 +86,7 @@ const chatWidgetStyles = makeStyles({ name: 'ylp-chat-widget' })((theme: Theme) 
     },
     chatWidgetFooter: {
       borderTop: `1px solid ${theme.palette.divider}`,
-      paddingBlock: theme.spacing(1)
+      paddingTop: theme.spacing(1)
     },
     chatWidgetEmojiPickerContainer: {
       position: 'absolute',
@@ -107,7 +108,7 @@ const chatWidgetStyles = makeStyles({ name: 'ylp-chat-widget' })((theme: Theme) 
       display: 'flex',
       alignItems: 'center',
       background: theme.palette.common.white,
-      paddingBlock: theme.spacing(1)
+      paddingTop: theme.spacing(1)
     },
     teamDropdown: {
       borderRadius: '28px',
@@ -227,158 +228,53 @@ function ChatWidget(props: Readonly<ChatWidgetProps>): JSX.Element {
   }
 
   return (
-    <>
-      <DataCard>
-        <Typography sx={{ fontWeight: 'bold' }}>
+    <DataCard>
+      <Box
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        sx={{ paddingBottom: theme.spacing(1) }}
+      >
+        <Typography sx={{ fontWeight: 'bold', paddingBottom: '8px' }}>
           {`${t('messages')} ${nbUnread > 0 ? `(+${nbUnread})` : ''}`}
         </Typography>
-        <Box position="relative">
-          <Dialog open={here} onClose={() => setHere(false)}>
-            <DialogContent>
-              <DialogContentText>
-                You asked for the beast, meet her at <Link href="https://eightsins.fr">www.eightsins.fr</Link>
-              </DialogContentText>
-            </DialogContent>
-          </Dialog>
-          <Box
-            ref={content}
-            className={classes.chatWidgetContent}
-            data-testid="chat-card-messages"
-          >
-            {messages.map((msg): JSX.Element => (
-              <ChatMessage
-                key={msg.id}
-                text={msg.text}
-                privateMsg={msg.private}
-                author={getUserName(msg.user.firstName, msg.user.lastName, msg.user.fullName)}
-                timestamp={msg.timestamp}
-                ack={msg.destAck}
-                isMine={msg.authorId === userId}
-              />
-            ))}
-          </Box>
-          {showPicker &&
-            <div
-              id="chat-widget-emoji-picker"
-              data-testid="chat-widget-emoji-picker"
-              className={classes.chatWidgetEmojiPickerContainer}
-              onKeyDown={onEmojiPickerKeyPress}
-            >
-              <EmojiPicker
-                width="100%"
-                height="440px"
-                lazyLoadEmojis
-                emojiStyle={EmojiStyle.NATIVE}
-                onEmojiClick={onEmojiClick}
-              />
-            </div>
-          }
-          <div id="chat-widget-footer" className={classes.chatWidgetFooter}>
-            {isUserHcp &&
-              <Tabs
-                className={classes.chatWidgetTabs}
-                value={inputTab}
-                onChange={handleChange}
-                sx={{ marginBottom: theme.spacing(1) }}
+        {isUserPatient &&
+          <FormGroup>
+            <Badge color="primary" variant="dot" invisible={hasNoUnreadMessages()}
+                   data-testid="unread-messages-badge">
+              <Select
+                data-testid="chat-widget-team-scope"
+                defaultValue={teamId}
+                IconComponent={KeyboardArrowDownIcon}
+                onChange={onTeamSelected}
+                variant="outlined"
+                className={classes.teamDropdown}
               >
-                <Tab
-                  className={classes.chatWidgetTab}
-                  label={t('chat-footer-reply')}
-                  aria-label={t('chat-footer-reply')}
-                  data-testid="chat-card-reply"
-                  onClick={() => {
-                    setPrivateMessage(false)
-                  }}
-                />
-                <Tab
-                  className={classes.chatWidgetTab}
-                  label={t('chat-footer-private')}
-                  aria-label={t('chat-footer-private')}
-                  data-testid="chat-card-private"
-                  onClick={() => {
-                    setPrivateMessage(true)
-                  }}
-                />
-              </Tabs>
-            }
-            <div ref={inputRow} className={classes.chatWidgetInputRow}>
-              <Button
-                id="chat-widget-emoji-button"
-                data-testid="chat-widget-emoji-button"
-                className={classes.iconButton}
-                onClick={() => {
-                  setShowPicker(true)
-                }}
-              >
-                <SentimentSatisfiedOutlinedIcon />
-              </Button>
-              <TextField
-                id="standard-multiline-flexible"
-                placeholder={t('chat-footer-start-writing')}
-                size="small"
-                multiline
-                maxRows={3}
-                value={inputText}
-                onChange={event => {
-                  checkThisOut(event.target.value)
-                  setInputText(event.target.value)
-                }}
-                InputLabelProps={{ shrink: false }}
-                data-testid="chat-card-input"
-              />
-              <Button
-                id="chat-widget-send-button"
-                disabled={inputText.length < 1}
-                className={classes.iconButton}
-                arial-label={t('send')}
-                title={t('send')}
-                data-testid="chat-card-send"
-                onClick={sendMessage}
-              >
-                <SendIcon />
-              </Button>
-            </div>
-          </div>
-        </Box>
-      </DataCard>
-    <GenericDashboardCard
-      title={`${t('messages')} ${nbUnread > 0 ? `(+${nbUnread})` : ''}`}
-      data-testid="chat-card"
-      action={isUserPatient &&
-        <FormGroup>
-          <Badge color="primary" variant="dot" invisible={hasNoUnreadMessages()} data-testid="unread-messages-badge">
-            <Select
-              data-testid="chat-widget-team-scope"
-              defaultValue={teamId}
-              IconComponent={KeyboardArrowDownIcon}
-              onChange={onTeamSelected}
-              variant="outlined"
-              className={classes.teamDropdown}
-            >
-              {
-                teams.map((team, index) =>
-                  <MenuItem
-                    key={index}
-                    data-testid={`chat-widget-team-scope-menu-${TeamUtils.formatTeamNameForTestId(team.name)}-option`}
-                    value={team.id}
-                  >
-                    {team.name}
-                    {unreadMessagesByTeamForPatient?.[team.id] &&
-                      <Badge
-                        variant="dot"
-                        color="primary"
-                        sx={{ p: 0.5 }}
-                        data-testid={`unread-messages-badge-team-${team.id}`}
-                      />
-                    }
-                  </MenuItem>
-                )
-              }
-            </Select>
-          </Badge>
-        </FormGroup>
-      }
-    >
+                {
+                  teams.map((team, index) =>
+                    <MenuItem
+                      key={index}
+                      data-testid={`chat-widget-team-scope-menu-${TeamUtils.formatTeamNameForTestId(team.name)}-option`}
+                      value={team.id}
+                    >
+                      {team.name}
+                      {unreadMessagesByTeamForPatient?.[team.id] &&
+                        <Badge
+                          variant="dot"
+                          color="primary"
+                          sx={{ p: 0.5 }}
+                          data-testid={`unread-messages-badge-team-${team.id}`}
+                        />
+                      }
+                    </MenuItem>
+                  )
+                }
+              </Select>
+            </Badge>
+          </FormGroup>
+        }
+      </Box>
+
       <Box position="relative">
         <Dialog open={here} onClose={() => setHere(false)}>
           <DialogContent>
@@ -487,8 +383,7 @@ function ChatWidget(props: Readonly<ChatWidgetProps>): JSX.Element {
           </div>
         </div>
       </Box>
-    </GenericDashboardCard>
-    </>
+    </DataCard>
   )
 }
