+++ packages/yourloops/components/dashboard-cards/models/chart-prefs.model.ts
