+++ packages/yourloops/components/dashboard-cards/patient-dashboard.tsx
@@ -41,12 +41,12 @@ import {
   RESPONSIVE_GRID_FULL_WIDTH,
   RESPONSIVE_GRID_HALF_WIDTH
 } from '../../css/css-utils'
-import MedicalFilesWidget from './medical-files/medical-files-widget'
+import MedicalFilesCard from './medical-files/medical-files-card'
 import MonitoringAlertsCard from '../monitoring-alert/monitoring-alerts-card'
 import { makeStyles } from 'tss-react/mui'
 import ChatWidget from '../chat/chat-widget'
 import { DEFAULT_DASHBOARD_TIME_RANGE_DAYS } from '../patient-data/patient-data.utils'
-import { DeviceUsageWidget } from './device-usage-widget'
+import { DevicesColumn } from './devices/devices-column'
 import { useParams } from 'react-router-dom'
 import TeamUtils from '../../lib/team/team.util'
 import { useTeam } from '../../lib/team'
@@ -154,7 +154,7 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
       </Grid>
 
       <Grid item xs={gridWidgetSize}>
-        <DeviceUsageWidget
+        <DevicesColumn
           dateFilter={dateFilter}
           medicalDataService={medicalDataService}
           patient={patient}
@@ -167,7 +167,7 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
             {user.isUserHcp() &&
               <MonitoringAlertsCard patient={patient} />
             }
-            <MedicalFilesWidget patient={patient} />
+            <MedicalFilesCard patient={patient} />
           </Grid>
 
           <Grid item xs={gridWidgetSize} className={classes.gridItemContainer}>
